pyclient: factor out ZethAddress and related code into its own file
