+++ pyClient/test_commands/mock.py
@@ -4,7 +4,8 @@
 #
 # SPDX-License-Identifier: LGPL-3.0+
 
-from zeth.mixer_client import EncryptionKeyPair, ZethAddress
+from zeth.zeth_address import ZethAddress
+from zeth.encryption import EncryptionKeyPair
 from zeth.ownership import gen_ownership_keypair
 from zeth.prover_client import ProverClient
 from zeth.utils import get_contracts_dir, get_private_key_from_bytes, \
