+++ pyClient/test_commands/test_ether_mixing.py
@@ -9,9 +9,10 @@ import zeth.contracts
 import zeth.merkle_tree
 import zeth.utils
 import zeth.zksnark
+from zeth.zeth_address import ZethAddressPriv
 from zeth.contracts import MixOutputEvents
 from zeth.encryption import EncryptionPublicKey
-from zeth.mixer_client import ZethAddressPriv, MixerClient
+from zeth.mixer_client import MixerClient
 from zeth.wallet import Wallet, ZethNoteDescription
 import test_commands.mock as mock
 import test_commands.scenario as scenario
